requirements.txt
pandas==2.3.3
openpyxl==3.1.5
selenium==4.41.0
google-api-python-client==2.166.0
google-auth==2.39.0
